+++ canopen_402_driver/include/canopen_402_driver/motor.hpp
@@ -148,8 +148,8 @@ public:
 class Mode
 {
 public:
-  const uint16_t mode_id_;
-  Mode(uint16_t id) : mode_id_(id) {}
+  const int8_t mode_id_;
+  Mode(int8_t id) : mode_id_(id) {}
   typedef WordAccessor<
     (1 << Command402::CW_Operation_mode_specific0) |
     (1 << Command402::CW_Operation_mode_specific1) |
@@ -170,7 +170,7 @@ class ModeTargetHelper : public Mode
   std::atomic<bool> has_target_;
 
 public:
-  ModeTargetHelper(uint16_t mode) : Mode(mode) {}
+  ModeTargetHelper(int8_t mode) : Mode(mode) {}
   bool hasTarget() { return has_target_; }
   T getTarget() { return target_; }
   virtual bool setTarget(const double & val)
@@ -417,9 +417,9 @@ public:
   }
 
   virtual bool setTarget(double val);
-  virtual bool enterModeAndWait(uint16_t mode);
-  virtual bool isModeSupported(uint16_t mode);
-  virtual uint16_t getMode();
+  virtual bool enterModeAndWait(int8_t mode);
+  virtual bool isModeSupported(int8_t mode);
+  virtual int8_t getMode();
   bool readState();
   void handleDiag();
   /**
@@ -487,7 +487,7 @@ public:
    * @return false
    */
   template <typename T, typename... Args>
-  bool registerMode(uint16_t mode, Args &&... args)
+  bool registerMode(int8_t mode, Args &&... args)
   {
     return mode_allocators_
       .insert(std::make_pair(
@@ -517,12 +517,12 @@ public:
   }
 
 private:
-  virtual bool isModeSupportedByDevice(uint16_t mode);
-  void registerMode(uint16_t id, const ModeSharedPtr & m);
+  virtual bool isModeSupportedByDevice(int8_t mode);
+  void registerMode(int8_t id, const ModeSharedPtr & m);
 
-  ModeSharedPtr allocMode(uint16_t mode);
+  ModeSharedPtr allocMode(int8_t mode);
 
-  bool switchMode(uint16_t mode);
+  bool switchMode(int8_t mode);
   bool switchState(const State402::InternalState & target);
 
   std::atomic<uint16_t> status_word_;
@@ -534,12 +534,12 @@ private:
   State402 state_handler_;
 
   std::mutex map_mutex_;
-  std::unordered_map<uint16_t, ModeSharedPtr> modes_;
+  std::unordered_map<int8_t, ModeSharedPtr> modes_;
   typedef std::function<void()> AllocFuncType;
-  std::unordered_map<uint16_t, AllocFuncType> mode_allocators_;
+  std::unordered_map<int8_t, AllocFuncType> mode_allocators_;
 
   ModeSharedPtr selected_mode_;
-  uint16_t mode_id_;
+  int8_t mode_id_;
   std::condition_variable mode_cond_;
   std::mutex mode_mutex_;
   const State402::InternalState switching_state_;
